% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ld_fnd.R
\name{add_ld_fnd}
\alias{add_ld_fnd}
\title{Add Locke Data as funder to DESCRIPTION}
\usage{
add_ld_fnd(path = getwd())
}
\arguments{
\item{path}{path to package}
}
\description{
Add Locke Data as funder to DESCRIPTION
}
