% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ld_labels.R
\name{use_ld_labels}
\alias{use_ld_labels}
\title{Use LD official labels}
\usage{
use_ld_labels()
}
\description{
Use LD official labels
}
\details{
Get a token https://help.github.com/articles/creating-a-personal-access-token-for-the-command-line/
 and save it via usethis::edit_r_environ()
}
