% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ld_contributing.R
\name{use_ld_github}
\alias{use_ld_github}
\title{Add CONTRIBUTING.md in .github}
\usage{
use_ld_github(project_name)
}
\arguments{
\item{project_name}{Project name}
}
\description{
Add CONTRIBUTING.md in .github
}
